refactor locations
